timer: rework timer_set() so it verifies requested ticks
This patch reworks cAVS timer_set() interface used
to write into timer's compare register.
This change takes care of ticks request which are
"past time". If such condition occur we set compare
register to current time so the latency is minimum.

+++ src/arch/host/include/arch/drivers/timer.h
@@ -21,8 +21,8 @@ static inline void arch_timer_unregister(struct timer *timer) {}
 static inline void arch_timer_enable(struct timer *timer) {}
 static inline void arch_timer_disable(struct timer *timer) {}
 static inline uint32_t arch_timer_get_system(struct timer *timer) {return 0; }
-static inline int arch_timer_set(struct timer *timer,
-	uint64_t ticks) {return 0; }
+static inline int64_t arch_timer_set(struct timer *timer,
+				     uint64_t ticks) {return 0; }
 static inline void arch_timer_clear(struct timer *timer) {}
 
 #endif /* __ARCH_DRIVERS_TIMER_H__ */
